requirements.txt
astroid==2.14.1
autopep8==2.0.1
colorama==0.4.6
dill==0.3.6
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysql-connector-python==8.0.31
platformdirs==2.6.2
protobuf==3.20.1
pycodestyle==2.10.0
pylint==2.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
wrapt==1.14.1
